README.adoc: update AppVeyor CI badge
The URL for AppVeyor CI badge is not scoped to `master` branch builds.

This means that all AppVeyor CI builds (including Pull Requests) can
affect the current AppVeyor build status reflected on the `README.adoc`.

Let's fix the erroneous URL for the AppVeyor CI badge to scope it to
`master` branch builds only.