README.adoc: update Travis CI badge URL
We renamed our GitHub organisation name from `cs2113-ay1819s2-t09-1` to
`CS2113-AY1819S2-T09-1`.

This resulted in a change of the project URL on Travis, which also
affected the Travis CI badge URL, causing the badge URL to point to an
unknown repository (`cs2113-ay1819s2-t09-1/main`).

Let's update the Travis CI badge URL so that it points to the correct
repository and show the correct build status again.