README.adoc: update Coveralls badge URL
We renamed our GitHub organisation name from `cs2113-ay1819s2-t09-1` to
`CS2113-AY1819S2-T09-1`.

This resulted in a change of the project URL on Coveralls, which also
affected the Coveralls badge URL, causing the badge URL to point to an
unknown repository (`cs2113-ay1819s2-t09-1/main`).

Let's update the Coveralls badge URL so that it points to the correct
repository and show the correct code coverage status again.