DeveloperGuide: add `find` module use case
The use case in the developer guide does not reflect most of our system
functional requirements.

A newer developer that wish to contribute would have a harder time to
ensure the proper behaviour between the system and the user.

Let’s add a use case to show a representation of a user’s interaction
with the system