UI: update `ModuleCard` and `ModuleListCard.fxml`
The details in the `ModuleListCard` sidebar are difficult to comprehend.
Example: Credits are displayed as is without any labels, and users may
be confused or might not remember what the value means or represents.

Let's update the UI to make the data presented in the `ModuleListCard`
easier for users to comprehend by
* labelling the `Credits` value displayed
* re-ordering some of details presented to make it look neater
* update `ModuleCardHandle#getCredits()` to remove the label added