test/CommandParserTestUtil: address `javac` linter warnings
`CommandParserTestUtil#assertParseSuccess(...)` and
`CommandParserTestUtil#assertParseFailure(...)` expects an instance of
the raw `Parser` object instead of the required `Parser<?>`.

As a result, the `javac` linter is unable to determine the generic class
at compile time, and thus it reported warnings of unchecked methods.

This may appear to be harmless, but it may cause runtime bugs in the
future.

Let's address the `javac` linter warnings by fixing the methods'
argument types in `CommandParserTestUtil` class.