Model: implement `Comparable` interface
Very often, there is a need to compare/sort attributes of `Module` and
`DegreePlanner`. Yet, these attributes (such as `Code`, `Year`,
`Semester`) do not implement the `Comparable` interface to facilitate
sorting of data.

Let's update the attributes of `Module` and `DegreePlanner` to implement
the `Comparable` interface to simplify and standardise the way we
compare the values.