Test/TestApp: fix `initPrefs(...)` implementation and broken tests
TestApp does not set the requirement category list JSON filepath and the
degree planner list JSON filepath in `initPrefs(...)`.

This causes `sandbox/pref_testing.json` to include the default values of
`degreePlannerListFilePath` and `requirementCategoryListFilePath` in
`UserPrefs` whenever `TestApp#initPrefs(...)` is used. Unfortunately,
this is incorrect, as TestApp uses save data files in the `sandbox`
directory!

Let's fix the erroneous `TestApp#initPrefs(...)` implementation and
update the broken unit tests to reflect the correct behaviour of the
application.