Model/module: make `Code` store uppercase string
Code should be made case-insensitive, as `cs1010` and `CS1010`
essentially refers to the same module code. We should also disallow
storing module codes in lowercase to maintain consistency throughout our
application (as some module codes will be in uppercase, some will be in
lowercase).

Let's update `Code` to make it case-insensitive, but store them in
uppercase. Additionally, let's update the unit tests to ensure that
case-insensitive `Code` works as expected.