Ui: update `GuiSettings#DEFAULT_WIDTH` and `MainWindow.fxml`
`GuiSettings#DEFAULT_WIDTH` does not match the minimum width of the root
element in `MainWindow.fxml`. Also, the `minWidth` of the Requirement
Categories panel needs to be increased, as the title of the panel can
become truncated if the width of the application window is minimum.

Let's update `GuiSettings#DEFAULT_WIDTH` to match `MainWindow.fxml`, and
let's also rescale the `minWidth` of all the main panels in
`MainWindow.fxml`.