Logic/commands: enhance logic checks for module management commands
`add` and `edit` commands have yet to be updated to properly handle
interactions with co-requisite/semesters constraints. This may cause
data integrity of the degree plan to be violated, which may cause the
application to function correctly.

Let's update the `add` and `edit` commands to properly handle
interactions with co-requisite/semesters, and update the tests to ensure
correct behaviour of the commands. Additionally, let's introduce
`StringUtil#joinStreamAsString(...)` helper method to reduce repetitive
code when converting a stream to a delimited string.